Add simple-host module
Not too sure about this one still...!

I don't want to commit a simple-host Helm chart in this repo, but also
its not right that we mostly use Helm for workload orchestration, but
use raw K8s manifests orchestrated and templated via Terraform for
simple-host.

Ideally there would be an open source simple-host Helm chart ... the
fact that this doesn't already exist is perhaps a clue that Helm is not
really group at composing small bits and pieces like this.  But perhaps
my opinion on this should change...