Thread node.name output through the wait Job
This ensures that anything depending on module.node.name waits for the
wait Job to complete.

This is essentially an implicit depend_on = [module.node...].

Should think about this a bit more ... what outputs do we want to thread
through the Job like this?  Is this better than just outright declaring
the depends_on?