Add EKS test cases
Also move to providing prereqs via boto, rather than Terraform.

Motiviation for this was that we need to use the SG ID in this SG test
case, but if we create the SG using Terraform this is a lot messier ...
we'd have to split this up into two separate Terraforms anyway???

Might be worth changing the others to use this approach too.

It also might be worth calling the Module directly, rather than having
what's now a useless wrapper ...