WIP refactor back to three-step apply
Unfortunately kubernetes provider is simply too limited to make two-step
apply work ... and this way has the advantage(?) of pulling bootstrap
out into a configuration, for easy swapping between actual examples.
And another advantage: workload layer is entirely environment
independent.

Layers:
- bootstrap: maximal common set of shared AWS infra.
- example/infra: extra infra required for a given example.  Includes
  some Kubernetes stuff (EBS CSI, multus, etc).
- example/workload: workload stuff (kubernetes, helm, etc).

The real kicker was that the K8s provider not only needs to know how to
access the cluster at apply time, it also needs CRDs present in the
cluster at apply time.  In particular net-attach-defs ... this basically
broke linux-pod-with-net-attach.

We could've moved this back to a Helm chart - maybe that's a reasonable
approach.  But there's also the question of the multus installation race
condition which is made much easier when multus is in a lower layer than
the workload.

Tbh I'm not really sure where this is headed.  I don't think this is
ideal for "lab" use necessarily - we're possibly at the point where maybe
Terraform is reaching its limits (already!) in that use case, and we
should be looking at cdktf, or alternatives (control plane based like
crossplane/nephio/osm model).  We're then in a bit of a weird place
with this where we want to focus on the customer examples use case, in
which case we need e2e tests of course, but don't want to base all our
testing on this ... ugh.