Fix kubelet node-labels handling
Allow users to pass node labels explicitly to the node module, rather
than implicitly passing the single label "name=%s".

This is necessary because now the name is uniquely generated (by the
Configuration), and so we either have to set the name label to a fixed
name (as the behaviour used to be), or update the Helm template.