Add version constraints to all UT Terraform configurations
This fixes an issue in terraform-provider-aws and moto - v1.23.0 of the
TF provider is incompatible with moto, because it pulls in a new version
of the AWS SDK for Go which doesn't like moto using strings in the
createdAt field in the response.  It wants an integer instead.

Also fix a regression in the Node UT.